+++ README.md
@@ -1,0 +1,7 @@
+- builds on venv
+- 
+
+# The SNAPE_ENV variable checks whether the current python environment is managed by snape
+# The SNAPE_ROOT variable contains the path to all snape environments (default: ~/.snape)
+# The SNAPE_VENV variable defines the name of the default environment (default: snape)
+# The SNAPE_PY variable points to the python script to execute as fallback
