+++ sh/snape.bash
@@ -17,7 +17,7 @@ if [ -z "$SNAPE_LOCAL_VENV" ]; then
 	export SNAPE_LOCAL_VENV=".venv"
 fi
 
-SNAPE_PY=$(dirname "${BASH_SOURCE[0]}")/../py/snape.py
+SNAPE_PY=$(realpath "$(dirname "${BASH_SOURCE[0]}")/../snape/run.py")
 SNAPE_PY_CMD="help --help -h new touch delete rm list setup status possess"
 
 mkdir -p "$SNAPE_ROOT"
