+++ snape/__init__.py
@@ -1,0 +1,5 @@
+from .run import main
+
+__all__ = [
+    "main"
+]
