+++ snape/cli/__init__.py
@@ -1,0 +1,9 @@
+from snape.cli.parser import parser, subcommands
+from snape.cli.main import main
+from snape.cli.commands import *
+
+__all__ = [
+    "parser",
+    "subcommands",
+    "main"
+]
