+++ snape/cli/commands/__init__.py
@@ -1,0 +1,17 @@
+from .delete import snape_delete
+from .help import snape_help
+from .new import snape_new
+from .possess import snape_possess
+from .setup import snape_setup_init, snape_setup_remove
+from .status import snape_status
+
+
+__all__ = [
+    "snape_delete",
+    "snape_help",
+    "snape_new",
+    "snape_possess",
+    "snape_setup_init",
+    "snape_setup_remove",
+    "snape_status"
+]
