+++ snape/util/__init__.py
@@ -1,0 +1,2 @@
+from snape.util.path import *
+from snape.util.io import *
